Move multiple functions into other script files
