+++ assets/timeHandler.js
@@ -1,0 +1,5 @@
+export function incrementSeconds() {
+  intervalIsUsed = true;
+  seconds += 1;
+  displayStats(statsText, seconds, userMistakesCount, userKeyTypeCount);
+}
